refactor: `SectionRepository` 를 추상화 작업
