+++ src/main/java/nextstep/line/infrastructure/SectionJpaRepository.java
@@ -1,0 +1,8 @@
+package nextstep.line.infrastructure;
+
+import nextstep.line.domain.Section;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SectionJpaRepository extends JpaRepository<Section, Long>, SectionRepository {
+
+}
