+++ src/main/java/nextstep/line/infrastructure/SectionRepository.java
@@ -1,8 +1,8 @@
 package nextstep.line.infrastructure;
 
+import java.util.List;
 import nextstep.line.domain.Section;
-import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SectionRepository extends JpaRepository<Section, Long> {
-
+public interface SectionRepository {
+    List<Section> findAll();
 }
